Adapt net.ts
