+++ src/net.ts
@@ -1,9 +1,12 @@
 import * as assert from 'assert';
+import * as crypto from 'crypto';
 import * as fs from 'fs';
 import fetch from 'node-fetch';
+import * as path from 'path';
 import * as stream from 'stream';
 import * as util from 'util';
 import * as vscode from 'vscode';
+import * as zlib from 'zlib';
 
 const pipeline = util.promisify(stream.pipeline);
 
@@ -62,24 +65,34 @@ export interface GithubRelease {
   }>;
 }
 
-export async function download(
-  downloadUrl: string,
-  destinationPath: string,
-  progressTitle: string,
-  { mode }: { mode?: number } = {},
-) {
+interface DownloadOpts {
+  progressTitle: string;
+  url: string;
+  dest: string;
+  mode?: number;
+  gunzip?: boolean;
+}
+
+export async function download(opts: DownloadOpts) {
+  // Put artifact into a temporary file (in the same dir for simplicity)
+  // to prevent partially downloaded files when user kills vscode
+  const dest = path.parse(opts.dest);
+  const randomHex = crypto.randomBytes(5).toString('hex');
+  const tempFile = path.join(dest.dir, `${dest.name}${randomHex}`);
+
   await vscode.window.withProgress(
     {
       location: vscode.ProgressLocation.Notification,
       cancellable: false,
-      title: progressTitle,
+      title: opts.progressTitle,
     },
     async (progress, _cancellationToken) => {
       let lastPercentage = 0;
       await downloadFile(
-        downloadUrl,
-        destinationPath,
-        mode,
+        opts.url,
+        tempFile,
+        opts.mode,
+        Boolean(opts.gunzip),
         (readBytes, totalBytes) => {
           const newPercentage = (readBytes / totalBytes) * 100;
           progress.report({
@@ -92,6 +105,8 @@ export async function download(
       );
     },
   );
+
+  return fs.promises.rename(tempFile, opts.dest);
 }
 
 /**
@@ -104,6 +119,7 @@ async function downloadFile(
   url: string,
   destFilePath: fs.PathLike,
   mode: number | undefined,
+  gunzip: boolean,
   onProgress: (readBytes: number, totalBytes: number) => void,
 ): Promise<void> {
   const res = await fetch(url);
@@ -136,13 +152,13 @@ async function downloadFile(
   });
 
   const destFileStream = fs.createWriteStream(destFilePath, { mode });
+  const srcStream = gunzip ? res.body.pipe(zlib.createGunzip()) : res.body;
 
-  await pipeline(res.body, destFileStream);
+  await pipeline(srcStream, destFileStream);
   return new Promise<void>(resolve => {
     destFileStream.on('close', resolve);
     destFileStream.destroy();
-
-    // Details on workaround: https://github.com/rust-analyzer/rust-analyzer/pull/3092#discussion_r378191131
-    // Issue at nodejs repo: https://github.com/nodejs/node/issues/31776
+    // This workaround is awaiting to be removed when vscode moves to newer nodejs version:
+    // https://github.com/rust-analyzer/rust-analyzer/issues/3167
   });
 }
